+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Added husky for running pre-commit hook.
 -   Run npm test on pre-commit hook.
 -   Added formatting on pre-commit hook.
+-   Refactored formatting
 
 ## v1.0.1 - (July 14, 2023)